requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
asynctest==0.13.0
atpublic==2.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.4
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
configobj==5.0.6
configparser==5.1.0
cryptography==36.0.0
dictdiffer==0.9.0
diskcache==5.3.0
distro==1.6.0
docker-pycreds==0.4.0
docopt==0.6.2
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.3
faiss-gpu==1.7.2
flatten-dict==0.4.2
flufl-lock==6.0
frozenlist==1.2.0
fsspec==2021.11.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.9
gitpython==3.1.24
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
grandalf==0.6
grpcio==1.42.0
httplib2==0.20.2
hyperopt==0.2.4
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
joblib==1.1.0
mailchecker==4.1.1
markdown==3.3.6
multidict==5.2.0
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
pathtools==0.1.2
phonenumbers==8.12.38
ply==3.11
promise==2.3
protobuf==3.19.1
psutil==5.8.0
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pydrive2==1.10.0
pygit2==1.7.1
pygments==2.10.0
pygtrie==2.4.2
pyopenssl==21.0.0
pyparsing==3.0.6
pytz==2021.3
pyyaml==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.15.2
rsa==4.8
ruamel-yaml==0.17.17
ruamel-yaml-clib==0.2.6
sentry-sdk==1.5.0
shortuuid==1.0.8
shtab==1.5.2
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
voluptuous==0.12.2
wandb
wcwidth==0.2.5
werkzeug==2.0.2
xmltodict==0.12.0
yarg==0.1.9
yarl==1.7.2
yaspin==2.1.0
zc-lockfile==2.0
zipp==3.6.0
